Pushed last edits on build and clean commands
