+++ src/main/java/ch/heigvd/statique/command/Clean.java
@@ -1,5 +1,6 @@
 package ch.heigvd.statique.command;
 
+import java.io.File;
 import java.util.concurrent.Callable;
 import picocli.CommandLine.Command;
 
@@ -7,8 +8,27 @@ import picocli.CommandLine.Command;
 public class Clean implements Callable<Integer> {
 
   @Override public Integer call() {
-    System.out.printf("clean");
+
+    String currentPath = System.getProperty("user.dir") + "/myStaticSite";
+    String buildPath = currentPath + "/build";
+
+    final File myStaticSite = new File(currentPath);
+    deleteDirectory(myStaticSite);
+
     return 1;
+  }
+
+
+
+
+  boolean deleteDirectory(File directoryToBeDeleted) {
+    File[] allContents = directoryToBeDeleted.listFiles();
+    if (allContents != null) {
+      for (File file : allContents) {
+        deleteDirectory(file);
+      }
+    }
+    return directoryToBeDeleted.delete();
   }
 
 }
